pyproject.toml
[project]
name = "deep-research"
version = "0.1.0"
description = "Deep Research planning workflows"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "llama-index-workflows>=2.13.0,<3.0.0",
    "python-dotenv>=1.1.0",
    "jsonref>=1.1.0",
    "click>=8.2.1,<8.3.0",
    "httpx>=0.28.1",
    "llama-index-core>=0.14.0",
    "llama-index-llms-openai>=0.6.6",
    "respx>=0.22.0,<1",
    "llama-cloud>=1.0.0b7",
    "oxylabs>=2.0.0",
    "llama-index-readers-oxylabs>=0.1.2",
    "llama-index-readers-web>=0.5.6",
    "trafilatura>=2.0.0",
]

[project.scripts]
deep-research = "deep_research.cli:main"

[dependency-groups]
dev = [
    "ruff>=0.11.10",
    "typescript>=0.0.12",
    "ty>=0.0.2",
    "pytest>=8.4.1",
    "hatch>=1.14.1",
    "pytest-asyncio>=1.3.0",
    "playwright>=1.56.0,<2",
    "psutil>=7.1.3",
    "pytest-playwright>=0.7.2",
    "pytest-timeout>=2.3.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.envs.default.scripts]
"format" = "ruff format ."
"format-check" = "ruff format --check ."
"lint" = "ruff check --fix ."
"lint-check" = ["ruff check ."]
typecheck = "ty check src"
test = "pytest"
"all-check" = ["format-check", "lint-check", "test"]
"all-fix" = ["format", "lint", "test"]

[tool.pytest.ini_options]
timeout = 120
timeout_method = "thread"

[tool.llamadeploy]
env_files = [".env"]
llama_cloud = true

[tool.llamadeploy.workflows]
research-metadata = "deep_research.research_metadata_workflow:workflow"
deep-research-plan = "deep_research.research_plan_workflow:workflow"
function-agent = "deep_research.function_agent_workflow:workflow"

[tool.llamadeploy.ui]
directory = "ui"
